% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rank_Predictor_Final.R
\name{Predict_Rank}
\alias{Predict_Rank}
\title{This was created by Devansh Gohil, IPM '24 Batch
If you need help, please email me on i24devanshg@iimidr.ac.in
Predict Academic Ranks Based on CGPA and Term}
\usage{
Predict_Rank(x, y = "Aggregate")
}
\arguments{
\item{x}{A numeric value representing CGPA (e.g., 3.1).}

\item{y}{A string specifying the term/year (e.g., "Year 1").}
}
\value{
A list containing predicted ranks, percentile, CGPA on a scale of 10, and other metrics.
}
\description{
This was created by Devansh Gohil, IPM '24 Batch
If you need help, please email me on i24devanshg@iimidr.ac.in
Predict Academic Ranks Based on CGPA and Term
}
\examples{
Predict_Rank(3.1, "Year 1")
}
